카드+보드 연동
