+++ front/src/components/auth/Login.js
@@ -23,7 +23,7 @@ const Login = () => {
 
         event.preventDefault();
 
-        fetch("http://54.180.109.210/auth/register", {
+        fetch("http://localhost:4000/auth/register", {
             method: "POST",
             headers: { "Content-Type": "application/json" },
             body: JSON.stringify({ email: Signemail.current.value, name: Signusername.current.value, password: Signpassword.current.value, passwordCheck: SignConfirmpassword.current.value })
@@ -41,7 +41,7 @@ const Login = () => {
 
     const loginsubmithandler = (event) => {
         event.preventDefault();
-        fetch("http://54.180.109.210/auth/login", { method: "POST", headers: { "Content-Type": "application/json" }, body: JSON.stringify({ email: loginemail.current.value, password: loginpassword.current.value }) })
+        fetch("http://localhost:4000/auth/login", { method: "POST", headers: { "Content-Type": "application/json" }, body: JSON.stringify({ email: loginemail.current.value, password: loginpassword.current.value }) })
             .then(res => res.json())
             .then(resData => {
                 session.setItem("access_token", resData.access_token)
