+++ front/src/components/board/Create_board.js
@@ -14,7 +14,7 @@ const Createboard = (props) => {
     const handleSubmit = () => {
         props.click()
         console.log(boardTitle, visibility, color)
-        fetch("http://54.180.109.210/board/", { method: "POST", headers: { "Content-Type": "application/json", "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImNobHhvZHVkMDRAbmF2ZXIuY29tIiwic3ViIjozLCJpYXQiOjE3MDQ2MjMyMTF9.V-lfby5HCDBl9BBK7rgHwRqDE-nh46HQ8G4RRebfS7Y" }, body: JSON.stringify({ name: boardTitle, description: visibility, color: color }) }).then(res => res.json()).then(resData => console.log(resData)).catch(err => console.log(err))
+        fetch("http://localhost:4000/board/", { method: "POST", headers: { "Content-Type": "application/json", "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImNobHhvZHVkMDRAbmF2ZXIuY29tIiwic3ViIjozLCJpYXQiOjE3MDQ2MjMyMTF9.V-lfby5HCDBl9BBK7rgHwRqDE-nh46HQ8G4RRebfS7Y" }, body: JSON.stringify({ name: boardTitle, description: visibility, color: color }) }).then(res => res.json()).then(resData => console.log(resData)).catch(err => console.log(err))
     };
 
     return (
